requirements.txt
asgiref==3.8.1
autopep8==2.3.1
Django==5.1.4
django-cors-headers==4.6.0
django-dirtyfields==1.9.5
django-dotenv==1.4.2
django-filter==24.3
django-silk==5.3.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
gprof2dot==2024.6.6
inflection==0.5.1
packaging==24.2
pillow==11.0.0
pycodestyle==2.12.1
PyJWT==2.10.1
python-magic==0.4.27
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
